make: add IPASIR_LIB and IPASIR_LIBS variable
When compiling, the variable IPASIR_LIB can be used to point to a library
of a SAT solver that implements the IPASIR interface. If that solver uses
additional variables, the variable IPASIR_LIBS can be specified as well
and is then used during linking.

The following command can be used to compile with an IPASIR solver. To
enable the CLI for solver selection, the tool has to be compiled in expert
mode (adding another xm=xp). The below command line assumes that the
variable IPASIR_LIB_FILE points to the ipasir archive file, and the
variable IPASIR_LIBS contains the linker flags required for the selected
solver to be linked.

make IPASIR_LIB=$IPASIR_LIB_FILE IPASIR_LIBS='$IPASIR_LIBS' xm=xp